MOve normalise_signal and denormaliased_signal to clarity.util.signal_processing
